+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- webapp: Fix toggle of selected media
 - webapp: Fix single view for not yet loaded media
 - webapp: Fix pre-loading of previews